book_store/requirement.txt
Django==1.11.6
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
mysqlclient==1.4.2.post1
pytz==2019.1
